postit server

+++ staff/aleix-caminal/postit-web-app/src/logic.js
@@ -1,12 +1,15 @@
-import { UsersTable, BoardsTable, PostsTable } from './model'
-import sha256 from 'js-sha256';
+const { UsersTable, BoardsTable, PostsTable } = require('./model')
+const { sha256 } = require('js-sha256')
 
 const LOGIC = {
     users: new UsersTable(),
     boards: new BoardsTable(),
     posts: new PostsTable(),
+    auth: JSON.parse(sessionStorage.getItem('auth')) || {},
 
     addBoard(form, user_id) {
+        if (typeof form !== 'object' || form.tagName !== 'FORM') throw Error('no form passed as argument')
+
         if (this.validate(form, ['title'])) {
             this.boards.newEntity({
                 title: form.querySelector('input[name="title"]').value,
@@ -25,6 +28,8 @@ const LOGIC = {
     },
 
     deleteBoard(id) {
+        if (!id) throw Error('id is not valid')
+
         const board = this.boards.get(id)
         board.delete()
         return this.boards.find({
@@ -33,6 +38,9 @@ const LOGIC = {
     },
 
     updateBoard(id, title) {
+        if (!id) throw Error('id is not valid')
+        if (!title) throw Error('title is not valid')
+
         const board = this.boards.get(id)
         board.title = title
         board.save()
@@ -42,6 +50,8 @@ const LOGIC = {
     },
 
     addPost(input, board_id) {
+        if (typeof input !== 'object' || input.tagName !== 'INPUT') throw Error('no input passed as argument')
+
         if (input) {
             this.posts.newEntity({
                 title: input.value,
@@ -56,6 +66,8 @@ const LOGIC = {
     },
 
     deletePost(id) {
+        if (!id) throw Error('id is not valid')
+
         const post = this.posts.get(id)
         post.delete()
         return this.posts.find({
@@ -63,14 +75,17 @@ const LOGIC = {
         })
     },
 
-    register(form) {
+    register(form, callback) {
+        if (typeof form !== 'object' || form.tagName !== 'FORM') throw Error('no form passed as argument')
+
         if (this.validate(form, ['name', 'username', 'password', 'confirm_password'])) {
             if (form.querySelector('input[name="password"]').value === form.querySelector('input[name="confirm_password"]').value) {
-                return this.users.newEntity({
+                this.users.newEntity({
                     name: form.querySelector('input[name="name"]').value,
                     username: form.querySelector('input[name="username"]').value,
                     password: sha256(form.querySelector('input[name="password"]').value)
-                }).save();
+                }).save()
+                callback()
             } else {
                 this.error('Passwords do not match');
                 form.querySelector('input[name="password"]').classList.add('is-invalid')
@@ -81,12 +96,15 @@ const LOGIC = {
         }
     },
 
-    login(form) {
+    login(form, callback) {
+        if (typeof form !== 'object' || form.tagName !== 'FORM') throw Error('no form passed as argument')
+
         if (this.validate(form, ['username', 'password'])) {
             let auth = this.findAuth(form.querySelector('input[name="username"]').value, sha256(form.querySelector('input[name="password"]').value))
             if (auth) {
                 sessionStorage.setItem('auth', JSON.stringify(auth))
-                return auth
+                this.auth = auth
+                callback(auth)
             } else {
                 this.error('Username or password are invalid')
                 form.querySelector('input[name="username"]').classList.add('is-invalid')
@@ -97,12 +115,16 @@ const LOGIC = {
         }
     },
 
-    logout() {
+    logout(callback) {
         sessionStorage.removeItem('auth')
-        return {}
+        this.auth = {}
+        callback()
     },
 
     findAuth(username, password) {
+        if (!username) throw Error('username is not valid')
+        if (!password) throw Error('password is not valid')
+
         try {
             const user_id = this.users.find({
                 username: username,
@@ -114,13 +136,13 @@ const LOGIC = {
         }
     },
 
-    getAuth() {
-        return JSON.parse(sessionStorage.getItem('auth')) || {}
+    isAuthenticated() {
+        return this.auth && Object.keys(this.auth).length > 0
     },
 
-    validate: function(form, inputs) {
-        if (typeof form !== 'object' || form.tagName !== 'FORM') throw Error('no form passed as argument');
-        if (!Array.isArray(inputs) || inputs.length < 1) throw Error('array is not valid');
+    validate(form, inputs) {
+        if (typeof form !== 'object' || form.tagName !== 'FORM') throw Error('no form passed as argument')
+        if (!Array.isArray(inputs) || inputs.length < 1) throw Error('array is not valid')
 
         var result = 1;
         for (var i in inputs) {
@@ -136,11 +158,11 @@ const LOGIC = {
         return result;
     },
 
-    error: function(message) {
-        if (!message) throw Error('message is not valid');
+    error(message) {
+        if (!message) throw Error('message is not valid')
 
-        alert(message);
+        alert(message)
     }
 }
 
-export default LOGIC
+module.exports = LOGIC
